--- lib/Shared/date_picker.dart
@@ -1,81 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter/cupertino.dart';
-
-class DatePickerDemo extends StatefulWidget {
-  @override
-  _DatePickerDemoState createState() => _DatePickerDemoState();
-}
-
-class _DatePickerDemoState extends State<DatePickerDemo> {
-  DateTime selectedDate = DateTime.now();
-
-  /// This decides which day will be enabled
-  /// This will be called every time while displaying day in calender.
-  bool _decideWhichDayToEnable(DateTime day) {
-    if ((day.isAfter(DateTime.now().subtract(Duration(days: 1))) &&
-        day.isBefore(DateTime.now().add(Duration(days: 10))))) {
-      return true;
-    }
-    return false;
-  }
-
-  _selectDate(BuildContext context) {
-    final ThemeData theme = Theme.of(context);
-    assert(theme.platform != null);
-    return buildMaterialDatePicker(context);
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Center(
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: <Widget>[
-            SizedBox(
-              height: 20.0,
-            ),
-            RaisedButton(
-              onPressed: () => _selectDate(context),
-              child: Text(
-                'Select date',
-                style:
-                    TextStyle(color: Colors.black, fontWeight: FontWeight.bold),
-              ),
-              color: Colors.greenAccent,
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  buildMaterialDatePicker(BuildContext context) async {
-    final DateTime picked = await showDatePicker(
-      context: context,
-      initialDate: selectedDate,
-      firstDate: DateTime(2000),
-      lastDate: DateTime(2025),
-      initialEntryMode: DatePickerEntryMode.calendar,
-      initialDatePickerMode: DatePickerMode.day,
-      selectableDayPredicate: _decideWhichDayToEnable,
-      helpText: 'Start Date',
-      cancelText: 'Cancel',
-      confirmText: 'Confirm',
-      errorFormatText: 'Enter valid date',
-      errorInvalidText: 'Enter date in valid range',
-      fieldLabelText: 'Booking date',
-      fieldHintText: 'Month/Date/Year',
-      builder: (context, child) {
-        return Theme(
-          data: ThemeData.light(),
-          child: child,
-        );
-      },
-    );
-    if (picked != null && picked != selectedDate)
-      setState(() {
-        selectedDate = picked;
-      });
-  }
-}
