+++ lib/screens/leave_screen.dart
@@ -2,9 +2,8 @@ import 'dart:ui';
 
 import 'package:flutter/material.dart';
 
-
 import '../Shared/calendar.dart';
-import '../Shared/date_picker.dart';
+
 
 class ApplyLeaveScreen extends StatefulWidget {
   @override
@@ -16,6 +15,54 @@ class _ApplyLeaveScreenState extends State<ApplyLeaveScreen> {
   bool _checked = false;
   String _chosenValue;
   String _chosenValue1;
+
+  DateTime selectedDate = DateTime.now();
+
+  /// This decides which day will be enabled
+  /// This will be called every time while displaying day in calender.
+  bool _decideWhichDayToEnable(DateTime day) {
+    if ((day.isAfter(DateTime.now().subtract(Duration(days: 1))) &&
+        day.isBefore(DateTime.now().add(Duration(days: 10))))) {
+      return true;
+    }
+    return false;
+  }
+
+  _selectDate(BuildContext context) {
+    final ThemeData theme = Theme.of(context);
+    assert(theme.platform != null);
+    return buildMaterialDatePicker(context);
+  }
+
+  buildMaterialDatePicker(BuildContext context) async {
+    final DateTime picked = await showDatePicker(
+      context: context,
+      initialDate: selectedDate,
+      firstDate: DateTime(2000),
+      lastDate: DateTime(2025),
+      initialEntryMode: DatePickerEntryMode.calendar,
+      initialDatePickerMode: DatePickerMode.day,
+      selectableDayPredicate: _decideWhichDayToEnable,
+      helpText: 'Start Date',
+      cancelText: 'Cancel',
+      confirmText: 'Confirm',
+      errorFormatText: 'Enter valid date',
+      errorInvalidText: 'Enter date in valid range',
+      fieldLabelText: 'Booking date',
+      fieldHintText: 'Month/Date/Year',
+      builder: (context, child) {
+        return Theme(
+          data: ThemeData.light(),
+          child: child,
+        );
+      },
+    );
+    if (picked != null && picked != selectedDate)
+      setState(() {
+        selectedDate = picked;
+      });
+  }
+
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
@@ -75,13 +122,7 @@ class _ApplyLeaveScreenState extends State<ApplyLeaveScreen> {
                 crossAxisAlignment: CrossAxisAlignment.start,
                 children: [
                   RaisedButton(
-                    onPressed: () {
-                      Navigator.push(
-                        context,
-                        MaterialPageRoute(
-                            builder: (context) => DatePickerDemo()),
-                      );
-                    },
+                    onPressed: () => _selectDate(context),
                     child: Text(
                       'From Date*',
                       style: TextStyle(
@@ -105,13 +146,7 @@ class _ApplyLeaveScreenState extends State<ApplyLeaveScreen> {
                 crossAxisAlignment: CrossAxisAlignment.start,
                 children: [
                   RaisedButton(
-                    onPressed: () {
-                      Navigator.push(
-                        context,
-                        MaterialPageRoute(
-                            builder: (context) => DatePickerDemo()),
-                      );
-                    },
+                    onPressed: () => _selectDate(context),
                     child: Text(
                       'To Date*',
                       style: TextStyle(
@@ -140,7 +175,6 @@ class _ApplyLeaveScreenState extends State<ApplyLeaveScreen> {
                 dropdownColor: Colors.grey[300],
 
                 items: <String>[
-  
                   'Cl/ Contigency Leave',
                   'Optional Holiday',
                   'Special Privelage Leave',
@@ -169,9 +203,7 @@ class _ApplyLeaveScreenState extends State<ApplyLeaveScreen> {
                   });
                 },
               ),
-              
             ),
-            
             CheckboxListTile(
               title: Text('Apply for Half-Day'),
               controlAffinity: ListTileControlAffinity.leading,
